requirements.txt
asgiref==3.5.0
Django==3.2.8
django-cors-headers==3.10.0
djangorestframework==3.12.4
djangorestframework-simplejwt==5.0.0
gunicorn==20.1.0
psycopg2==2.9.3
PyJWT==2.3.0
python-dotenv==0.20.0
pytz==2022.1
sqlparse==0.4.2
